refactor: 배팅 금액 및 수익 계산 책임 Player로 변경
